commited

+++ post/views.py
@@ -9,24 +9,41 @@ from django.contrib.auth.mixins import LoginRequiredMixin
 
 
 # Create your views here.
-class PostListView(View,LoginRequiredMixin):
+class PostListView(LoginRequiredMixin,View):
     def get(self,request,*args,**kwargs):
-        posts=Post.objects.filter(user=request.user)
-        paginator=Paginator(posts,8)
-        page=request.GET.get("page",1)
-        page_obj=paginator.get_page(page)
-        return render(request,"post/list.html",context={"posts":page_obj})
+        try:
+           posts=Post.objects.filter(user=request.user)
+           paginator=Paginator(posts,8)
+           page=request.GET.get("page",1)
+           page_obj=paginator.get_page(page)
+           return render(request,"post/list.html",context={"posts":page_obj})
+        except Post.DoesNotExist:
+            return redirect(reverse("404 error"))
+        except Exception:
+            return redirect(reverse("500 error"))
 
 class PostDetailView(View):
     def get(self,request,id,*args,**kwargs):
-        post=Post.objects.get(id=id)
-        return render(request,"post/detail",context={"posts":post})
+        try:
+           post=Post.objects.get(id=id)
+           return render(request,"post/detail",context={"posts":post})
+        except Post.DoesNotExist:
+            return redirect(reverse("404 error"))
+        except Exception:
+            return redirect(reverse("500 error"))
 
-class PostEditView(View):
+class PostEditView(LoginRequiredMixin,View):
     def get(self,request,id,*args,**kwargs):
-        post=Post.objects.get(id=id)
-        form=PostForm(instance=post)
-        return render(request,"post/edit.html",context={"form":form})
+        try:
+            post=Post.objects.get(id=id)
+            #if post.user != request.user:
+            #    return redirect(reverse("404_error"))
+            form=PostForm(instance=post)
+            return render(request,"post/edit.html",context={"form":form})
+        except Post.DoesNotExist:
+            return redirect(reverse("404 error"))
+        except Exception:
+            return redirect(reverse("500 error"))
     def post(self,request,id,*args,**kwargs):
         post=Post.objects.get(id=id)
         form=PostForm(instance=post,data=request.POST,files=request.FILES)
@@ -61,17 +78,27 @@ def post_create(request):
             return redirect(reverse('post-list'))
         return render(request,"post/create.html",context={'form':form})
         
-class PostCreateView(View):
+class PostCreateView(LoginRequiredMixin,View):
     def get(self,request,id,*args,**kwargs):
-        form=PostForm()
-        return render(request,"post/create.html",context={'form':form})
+        try:
+           form=PostForm()
+           return render(request,"post/create.html",context={'form':form})
+        except Post.DoesNotExist:
+            return redirect(reverse("404 error"))
+        except Exception:
+            return redirect(reverse("500 error"))
     def post(self,request,id,*args,**kwargs):
-        print("POST DATA:", request.POST,request.FILES)
-        form=PostForm(request.POST,request.FILES)
-        if form.is_valid():
-            form.save()
-            return redirect(reverse('post-list'))
-        return render(request,"post/create.html",context={'form':form})
+        try:
+           print("POST DATA:", request.POST,request.FILES)
+           form=PostForm(request.POST,request.FILES)
+           if form.is_valid():
+               form.save()
+               return redirect(reverse('post-list'))
+           return render(request,"post/create.html",context={'form':form})
+        except Post.DoesNotExist:
+            return redirect(reverse("404 error"))
+        except Exception:
+            return redirect(reverse("500 error"))
    
 
 
@@ -85,16 +112,23 @@ def post_edit(request,id):
         post.save()
         return redirect(reverse('post-list'))
 
-class PostDeleteView(View):
+class PostDeleteView(LoginRequiredMixin,View):
     def post(self,request,*args,**kwargs):
-        post=Post.objects.get(id=request.POST.get("id"))
-        post.delete()
-        return redirect(reverse('post-list'))
+        try:
+           post=Post.objects.get(id=request.POST.get("id"))
+           post.delete()
+           return redirect(reverse('post-list'))
+        except Post.DoesNotExist:
+            return redirect(reverse("404 error"))
+        except Exception:
+            return redirect(reverse("500 error"))
    
 
 def post_delete(request):
     if request.method == "POST":
         post=Post.objects.get(id=request.POST.get("id"))
+        if post.user != request.user:
+            return redirect(reverse("404_error"))
         post.delete()
         return redirect(reverse('post-list'))
 
@@ -104,8 +138,13 @@ def post_detail(request,id):
 
 class IndexView(View):
     def get(self,request,*args,**kwargs):
-        posts=Post.objects.all()
-        paginator=Paginator(posts,8)
-        page=request.GET.get("page",1)
-        page_obj=paginator.get_page(page)
-        return render(request,"index.html",context={"posts":page_obj})
+        try:
+           posts=Post.objects.all()
+           paginator=Paginator(posts,8)
+           page=request.GET.get("page",1)
+           page_obj=paginator.get_page(page)
+           return render(request,"index.html",context={"posts":page_obj})
+        except Post.DoesNotExist:
+            return redirect(reverse("404 error"))
+        except Exception:
+            return redirect(reverse("500 error"))
